DPLT-1121: Generate and Support Strongly Typed Objects for DB Methods (#193)
This commit adds functionality for generating TypeScript interfaces and
a context object which reflect rows in each table defined in the SQL
table and respective functions for insert, select, upsert, update, and
delete which all take in a strongly formatted object. This allows for
autocomplete and type assertions for writing objects or calling
functions related to each defined table.

This is accomplished by using the library
[node-sql-parser](https://www.npmjs.com/package/node-sql-parser) which
contains functions for converting valid Postgres DDL into an AST or list
of tables.

In the frontend, I have added a button which triggers code gen and
attaches the types to Monaco. In addition, I set up listeners on
monaco's mounting status as well as on the schemaTypes state variable
which I introduced to store the generated types. These typesa re also
saved to localStorage to be retrieved. The following actions trigger a
code generation to take place: clicking the code gen button (Next to
format button), switching to the indexingLogic.js tab, when monaco
mounts, and when the schema and code are retrieved from local storage.
The AST is parsed using the new PgSchemaTypeGen class. The end result is
a context object formatted like so: context.db.TableName.methodName({}).

NOTE: The context object added to monaco is done so as a global
variable. Currently our indexers pass in a context parameter. This
causes the local parameter to shadow the global one, which prevents
autocomplete or type checking. In order to use the global one, the
context object must be removed from the function declaration for
getBlock.

In the backend, the context object received a similar refactor to the
context.db.TableName.methodName format. It also uses the parser library
but solely to get the table names to generate the context object.