refactor: Aggregate worker metrics in background (#355)
Previously, we would need to manually send the metric data to the main
thread, then write the metric there. This PR updates worker metrics so
it is done seamlessly in the background, metrics can be written as
normal regardless of whether in the main thread or not. This is achieved
by sending all metrics collected within a worker to the main thread, and
then aggregating them there.