feat: Cancel historical backfill process before starting anew (#363)
This PR updates Coordinator such that **only one** Historical Backfill
process can exist for a given Indexer Function. Currently, these
processes do not have awareness of each other, meaning a developer can
trigger multiple simultaneous backfills, with them all writing to the
same Redis Stream. The major changes are detailed below.

## Streamer/Historical backfill tasks
In this PR, I've introduced the `historical_block_processing::Streamer`
struct. This is used to encapsulate the async task which is spawned for
handling the Historical Backfill. I went with "Streamer", as I envision
this becoming the single process which handles both historical/real-time
messages for a given Indexer, as described in
https://github.com/near/queryapi/issues/216#issuecomment-1777938653.

`Streamer` is responsible for starting the `async` task, and provides a
basic API for interacting with it. Right now, that only includes
cancelling it. Cancelling a `Streamer` will drop the existing future,
preventing any new messages from being added, and then delete the Redis
Stream removing all existing messages.

## Indexer Registry Mutex/Async
`indexer_registry` is the in-memory representation of the given Indexer
Registry. Currently, we pass around the `MutexGuard` to mutate this
data, resulting in the lock being held for much longer than is needed,
blocking all other consumers from taking the lock. There isn't much
contention for this lock, so it isn't a problem right now, but could end
up being one in future.

Adding in the `Streamer` `Mutex` made mitigating this problem trivial,
so I went ahead and made some changes. I've updated the code so that we
pass around the `Mutex` itself, rather than the guard, allowing us to
the hold the lock for the least amount of time possible. This required
updating most methods within `indexer_registry` to `async`, so that we
could take the `async` `lock`.

With the `Mutex` being used rather than the `MutexGuard`, it seemed
sensible to add `indexer_registry` to `QueryAPIContext`, rather than
passing it round individually.