feat: Pre-Fetch Streamer Messages (#269)
Historical streamer messages are not fetched in coordinator prior to the
IndexerRunner call. As a result, we cannot apply the latency saving
benefits of having coordinator cache the streamer message for use by
runner. Instead, we want to pre fetch from S3 so that runner invocations
don't have to wait for the streamer message to be retrieved from S3.

In addition, it's possible for real-time messages to backup temporarily
preventing the cached message from being used. So, we also want to
prefetch any messages which aren't found in the cache.

The new workflow works by having two loops for each worker thread: a
producer and a consumer. The producer loads a promise for fetching the
block (either from cache or S3) into an array. The consumer then removes
the first element from the array and processes it, deleting the streamer
message upon success. While one block is being processed, the other
blocks are being fetched. This ensures that wait time is minimal. The
producer loop attempts to keep the array as close to full as possible.