refactor: Speed up Historical Backfill process (#379)
## Problem

The Historical backfill is composed of two main steps:
1. Fetching matching blocks from `near-delta-lake`/Databricks index
files
2. Manually filtering all blocks between the `last_indexed_block` from
`near-delta-lake` and the `block_height` in which the historical process
was triggered

Only after each of these steps are completed are the blocks flushed to
Redis. This leads to a large delay from when the process was triggered,
to when the blocks are actually executed by Runner, creating the
appearance of things 'not working'.

## Changes
This PR makes the following changes to reduce the time between trigger,
and execution:

### Flush blocks indexed blocks immediately
Fetching blocks from the index file (Step 1.), is relatively quick
process. Rather than wait for (Step 2.), we can flush the blocks
immediately, and then continue on to the following step.

### Prefetch blocks in manual filtering
In manual filtering, blocks are processed sequentially. To speed this
process up, we can fetch ahead so that we minimise the time spent
waiting for S3 requests. Luckily, `near-lake-framework` does exactly
this, therefore manual filtering has been refactored to use this.

## Results
The following is from a backfill of 19,959,790 blocks run locally. I'd
expect the Before to be much lower given the geographical distance
compared to actual our infrastructure, but the results are still
positive :).

Time till first block appears on Redis:
- Before: 33 minutes
- After: 2 seconds

Time to completion:
- Before: 33 minutes
- After: 45 seconds