fix: Ensure Indexer Config Updates Are Read in Runner (#384)
Runner sets Indexer Config when the thread is started. As a result, it
does not react to updates to that config, such as updates to code. This
is a problem as that means unless runner is restarted, published code
changes won't be used. I've changed it so that the config is read each
iteration of the loop. That way, config updates will be consumed. In the
short term, this can be tuned such as reading every X loops and on every
failure, if need be. In the long term, improved communication between
coordinator and runner can facilitate coordinator communicating to
runner to read the config as opposed to doing so all the time.

In addition, the metrics for block wait duration and overall execution
duration were wrong. I've moved the start time to the correct spot.