fix: Try to resolve heap memory errors (#435) (#436)
We are seeing errors in prod relating to javascript running out of
memory due to the heap growing. Most likely culprit is the block
prefetch. There may be some characteristics in prod that cause this
error, but not in dev. Failures cause the entire machine to crash and
restart.

We'll reduce the prefetch queue size, which is likely the largest memory
eater in the worker, to hopefully avoid this worker heap memory issue.