feat: Create initial Block Streamer service (#428)
This PR introduces a new Rust service called Block Streamer. This is
essentially a refactored version of Historical Backfill, which does not
stop the 'manual filtering' portion of the backfill, continuing
indefinitely. Eventually, this service will be able to serve requests,
allowing control over these Block Streams. I've outlined the major
changes below.

## `trait S3Client`
Low level S3 requests have been abstracted behind a `trait`. This allows
us to use the `trait` as an argument, allowing the injection of mock
implementations in tests. We no longer need to use real access keys, and
can make more concrete assertions.

## `DeltaLakeClient`
The majority of the S3 related logic, `queryapi_coordinator/src/s3.rs`
included, has been refactored in to `DeltaLakeClient`. This client
encapsulates all logic relating the interaction/consumption of
`near-delta-lake` S3. Now, only a single method needs to be called from
`BlockStream` in order to get the relevant block heights for a given
contract_id/pattern. There's been a fair amount of refactor across all
the methods themselves, I've added tests to ensure the still behave as
expected.

## `BlockStream`
The refactored version of Historical Backfill, not much has changed here
in this version, the main change is that it now used `DeltaLakeClient`.
This will eventually be expanded to provide more control over the
Stream.

## `indexer_rules_*` & `storage`
Currently these exist as separate crates within the `indexer/`
workspace. Rather than creating a workspace for `block-streamer` I've
added the respective files for each to the single crate. This is
probably fine for `storage`, now called `redis`, but given
`indexer_rules_*` is also used in the Registry Contract, I'll probably
extract this in a later PR, to avoid it being duplicated.