Fix type gen bug (#449)
We have an issue where if the formatting of SQL is not current or we
encounter a case in which we do not support the type generation library
for the context method, we fail very abruptly.

This PR handles these errors better. 

- If sql/code encounters syntax error and formatting fails, we simply
return the unformatted code to the editor so the user can fix it.
- If code + sql both have syntax errors, the user will now see two
different alerts. One for each one of them.
- We throw the error, but we also log everything to console for
debugging purposes.
- We catch and throw type generation errors in a new variable.