feat: Runner gRPC endpoint (#446)
As part of implementing a new control plane to oversee QueryApi
resources, Runner needed endpoints with which control plane can send
commands. This PR adds code to create a new gRPC server with endpoints
to start, stop, or list Runner executors. This provides the control
plane the ability to fully control Runner, and removes the need for
bilateral decision making, which was a problem of the previous design.

The changes will be changed further as needs become more concrete,
closer to integration/release.