feat: Add created/updated at block heights to registry (#458)
This PR adds the `created_at_block_height` and `updated_at_block_height`
fields to `IndexerConfig` within the registry contract. The motive
behind this is to provide Coordinator V2 with a way for comparing the
actual and desired states of the system, i.e. if there is a mismatch
between the registry and the system, action should be taken. Without
versions, there is no way of making this comparison.

## Compilation Errors
~~I ran in to several issues trying to compile the `wasm32` binary, and
have outlined all these issues in
https://github.com/near/near-sdk-rs/issues/1125, as well as in the
`README.md` so that the fixes are documented. These fixes are a bit
janky, but I've tested the deployed contract and all seems to be ok.~~

These have been resolved, see:
https://github.com/near/queryapi/pull/458#issuecomment-1874977342

## Account Roles Migration
I've also included `account_roles` in this migration as we have some
incorrect accounts as `Owner`s (`pavelnear.near`). All owners will be
wiped and re-written from the contract default state. All `User`s will
remain.

## Coordinator V1
Coordinator V1 has been tested to ensure that it can still parse the
registry after these new fields have been applied.