feat: Create rust GRPC client for Runner (#491)
This PR creates a rust based GRPC client for Runner, for use within
Coordinator V2. For now, this exists as its own crate within the
top-level directory. The Coordinator PR was becoming quite large so I
decided to separate this out.

Additionally, I've made a couple changes to the Runner proto:
- Rename the package: `spec` -> `runner`
- Remove `executorId` from `StartExecutorRequest` - it should be
deterministic and assigned internally
- Add `version` to the executor - this will be used to determine whether
an executor should be restarted