Improve code/schema validation before registering (#495)
More info about this can be found on this
[issue](https://github.com/near/queryapi/issues/483)

Changes on this PR:
- Validate code & schema before registering the indexer
Logic: If formatting either the code or the schema fails, the `Publish`
button is disabled. Only if type generation errors or no errors are
detected, the `Publish` button will be enabled. More about this on this
[discussion](https://github.com/near/queryapi/pull/480#issuecomment-1878884650)


https://github.com/near/queryapi/assets/15988846/68f89cb4-f561-4e8a-9fa7-81c1a38e548c

Additionally 
- Created a reusable Modal with a global context to manage it, so we can
trigger it from any component to display some info
- Updated schema with granular error types for improved clarity
- Created a custom error class to filter by type Error. We can add more
fields on it if needed

---------

Co-authored-by: Juan Luis Santana <juanluis@near.org>
Co-authored-by: Roshaan Siddiqui <siddiqui.roshaan@gmail.com>