feat: Capture errors thrown within Coordinator (#515)
Currently, errors thrown within Coordinator V2 will bubble up to
`main()` and cause the entire application to exit. This PR captures
those errors, handling them accordingly.

Errors are handled in either of the following ways:
1. Exponentially retry - data which is 'critical' to the control, i.e.
the indexer registry, executor/stream list, will be continuously
retried, blocking the control loop from further progress as it would not
make sense to continue without this information.
2. Swallowed - actions such as starting/stopping executors/streams will
be logged and swallowed. This is preferable over exponential retries as
individual failures will not block the progress of the control loop,
therefore allowing other indexers to be acted on. Skipping should be
fine in this case as it will be retried in the next loop.

I expect this behaviour to evolve over time as we learn more about the
system, the important thing here is that Coordinator will not crash on
errors.