fix: Prevent skipping blocks from Redis Stream (#558)
`xread` returns messages _after_ the provided ID, there's no need to
increment the ID ourselves. If we increment to an ID that actually
exists within the stream, the message will be skipped permanently. e.g.
we have messages `1-0` and `1-1`, on first read we get `1-0`, we
increment to `1-1`, then on the second read we get nothing.