feat: Support Star Contract Filter in Block Streamer (#572)
Block Streamer will fail to process a * contract filter as it will try
to fetch index files for all account IDs. This is not only extremely
inefficient since we aren't planning to skip anything, but will also
fail as we limit the number of accounts fetched as a sanity check.

Block Streamer, if the filter is "*" specifically, now skips fetching
index files and begins reading from Lake directly instead.

This retains the behavior of block streamer reading block data before
pushing the height to Redis. There is an intended migration of Runner's
prefetch into Block Streamer. This change works well towards that goal.
However, obviously there is double work performed here: We fetch the
block, only to just put the height into Redis. If, in the short term and
before we migrate prefetch from Runner to Block Streamer, we need to
speed things up, we can list files form S3 instead. However, this work
would be removed after the mentioned migration of work.